more refactoring to remove circular dependency
