+++ smrt-server-analysis/src/main/scala/com/pacbio/secondary/smrtserver/client/ServiceAccessLayer.scala
@@ -51,8 +51,55 @@ class AnalysisServiceAccessLayer(baseUrl: URL)(implicit actorSystem: ActorSystem
     val ROOT_PT = "/secondary-analysis/resolved-pipeline-templates"
   }
 
+  def getReportPipeline: HttpRequest => Future[Report] = sendReceive ~> unmarshal[Report]
+  def getJobPipeline: HttpRequest => Future[EngineJob] = sendReceive ~> unmarshal[EngineJob]
+  // XXX this fails when createdBy is an object instead of a string
+  def getJobsPipeline: HttpRequest => Future[Seq[EngineJob]] = sendReceive ~> unmarshal[Seq[EngineJob]]
   def runJobPipeline: HttpRequest => Future[EngineJob] = sendReceive ~> unmarshal[EngineJob]
   def getPipelineTemplatePipeline: HttpRequest => Future[PipelineTemplate] = sendReceive ~> unmarshal[PipelineTemplate]
+
+  protected def getJobsByType(jobType: String): Future[Seq[EngineJob]] = getJobsPipeline {
+    Get(toUrl(ServiceEndpoints.ROOT_JOBS + "/" + jobType))
+  }
+
+  def getAnalysisJobs: Future[Seq[EngineJob]] = {
+    getJobsByType(JobTypes.PB_PIPE)
+  }
+
+  def getImportJobs: Future[Seq[EngineJob]] = {
+    getJobsByType(JobTypes.IMPORT_DS)
+  }
+
+  def getMergeJobs: Future[Seq[EngineJob]] = {
+    getJobsByType(JobTypes.MERGE_DS)
+  }
+
+  def getFastaConvertJobs: Future[Seq[EngineJob]] = {
+    getJobsByType(JobTypes.CONVERT_FASTA)
+  }
+
+  def getJobByAny(jobId: Either[Int, UUID]): Future[EngineJob] = {
+    jobId match {
+      case Left(x) => getJobById(x)
+      case Right(x) => getJobByUuid(x)
+    }
+  }
+
+  def getJobById(jobId: Int): Future[EngineJob] = getJobPipeline {
+    Get(toUrl(ServiceEndpoints.ROOT_JOBS + "/" + jobId))
+  }
+
+  def getJobByUuid(jobId: UUID): Future[EngineJob] = getJobPipeline {
+    Get(toUrl(ServiceEndpoints.ROOT_JOBS + "/" + jobId))
+  }
+
+  def getJobByTypeAndId(jobType: String, jobId: Int): Future[EngineJob] = getJobPipeline {
+    Get(toJobUrl(jobType, jobId))
+  }
+
+  def getAnalysisJobById(jobId: Int): Future[EngineJob] = {
+    getJobByTypeAndId(JobTypes.PB_PIPE, jobId)
+  }
 
   def importDataSet(path: String, dsMetaType: String): Future[EngineJob] = runJobPipeline {
     Post(
